ml-service/requirements.txt
fastapi
uvicorn[standard]
opencv-python
numpy
python-dotenv
loguru
pydantic